Refactor texto
El trozo de código que añade texto a las imágenes lo he separado
Había lógica redundante que aun me sorprende que funcionara
